src/requirements.txt
python-dotenv
agno
sqlalchemy
mcp
google-genai
PyMuPDF
python-docx
openpyxl
pytesseract
pandas
asyncio
fastapi
uvicorn
pydantic
psycopg-pool
firecrawl-py
psycopg2-binary
psycopg[binary]
pymongo
chromadb
langchain
langchain-community
langchain-chroma
langchain-text-splitters
pypdf
docx2txt
unstructured[all-docs]
sentence-transformers
langchain_huggingface
matplotlib